+++ frontend/src/common/constants/career.ts
