+++ frontend/src/common/types/userInfoResponse.ts
