+++ frontend/src/common/utils/careerValidator.ts
