+++ frontend/src/common/utils/priceValidator.ts
